+++ crates/broker/src/util/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod ansi;
+pub(crate) mod fs;
 pub(crate) mod terminal;
 pub(crate) mod utf8_stream;
 pub(crate) mod version;